requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
streamlit
python-docx
pdfplumber
plotly
streamlit_javascript
redis
pymongo
google-genai
pytest
fakeredis